refactoring types  and altering routes
