--- src/app/TFRRoute.ts
@@ -1,5 +1,0 @@
-import { Route } from '@angular/router';
-
-export interface TFRRoute extends Route {
-  label?: String;
-}
