+++ src/app/components/chip/chip.component.html
@@ -3,6 +3,6 @@
   selectable="true"
   (removed)="remove($event)"
   (click)="click($event)"
-  >{{ value.start_date | date : "shortDate" }}
+  >{{ value.description }}
   <mat-icon matChipRemove>cancel</mat-icon>
 </mat-chip>
