+++ src/app/components/tfr/tfr.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ApiService } from 'src/app/services/api.service';
 import { TfrManagementService } from 'src/app/services/tfr-management/tfr-management.service';
-import { Project } from 'src/app/types/types';
+import { Project } from 'src/app/shared/interfaces';
 
 @Component({
   selector: 'app-tfr',
@@ -24,8 +24,8 @@ export class TfrComponent implements OnInit {
     this.TfrId = Number(this.route.snapshot.paramMap.get('id'));
 
     /*
-      Error validation for the path variable. 
-      The path variable (the project_id) is expected to be a number. 
+      Error validation for the path variable.
+      The path variable (the project_id) is expected to be a number.
     */
     if (!Number.isInteger(this.TfrId)) {
       this.router.navigate(['/home']);
@@ -36,7 +36,7 @@ export class TfrComponent implements OnInit {
 
       /*
         The data that will be rendered in the screen is pre-fetched before the component
-        is loaded. This component has a resolver (refer to /services/project-resolver) that 
+        is loaded. This component has a resolver (refer to /services/project-resolver) that
         fetches the project to be displayed.
       */
       this.route.data.subscribe((response) => {
@@ -57,7 +57,7 @@ export class TfrComponent implements OnInit {
   }
 
   /*
-    Takes the user to the URL that enable him to edit the current 
+    Takes the user to the URL that enable him to edit the current
     project being displayed.
   */
   redirectToEditTfr() {
