+++ src/app/shared/interfaces.ts
@@ -1,3 +1,5 @@
+import { Route } from '@angular/router';
+
 export interface ProjectDTO {
   id?: number;
   name: string;
@@ -12,6 +14,39 @@ export interface ProjectDTO {
   created_at: Date;
   modified_by: number;
   modified_at: Date;
+}
+
+export interface Project {
+  id: number;
+  name: string;
+  vendor_id: number;
+  start_date: Date;
+  end_date: Date;
+  status: string;
+  version: number;
+  vendor_specific: { [key: string]: string };
+  is_deleted: Boolean;
+  created_by: number;
+  modified_by: number;
+  created_at: Date;
+  modified_at: Date;
+  milestones: Milestone[];
+  project_resources: ProjectResourceDTO[];
+}
+
+export interface Milestone {
+  id: number;
+  project_id: number;
+  description: string;
+  start_date: Date;
+  delivery_date: Date;
+  acceptance_date: Date;
+  is_deleted: Boolean;
+  tracker?: TrackerDTO;
+}
+
+export interface TFRRoute extends Route {
+  label?: String;
 }
 
 export interface TaskDTO {
